feat:WinningLotto organize
