+++ src/main/java/domain/LottoCorrect.java
@@ -1,0 +1,11 @@
+package domain;
+
+import lotto.Lotto;
+
+import java.util.List;
+
+public class LottoCorrect {
+
+    //입력 내용과 비교하기
+    public List<Integer> compare(LottoGenerate lottoGenerate, List<Lotto> lottos) {}
+}
