+++ src/main/java/util/WinningNumberParser.java
@@ -1,13 +1,12 @@
-package domain;
+package util;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
 public class WinningNumberParser {
 
-    public List<Integer> parse(String Number){
+    public static List<Integer> parse(String Number){
         //문자열을 숫자 리스트로 반환
         List<Integer> winningNumbers = Arrays.stream(Number.split(","))
                 .map(String::trim)
